# Various test speed ups