--- a/.circle/install-and-run-mongodb.sh
+++ b/.circle/install-and-run-mongodb.sh
@@ -19,8 +19,9 @@ wget -q ${DOWNLOAD_URL} -O /tmp/mongodb.tgz
 tar -xvf /tmp/mongodb.tgz -C ${MONGODB_DIR} --strip=1
 
 echo "Starting MongoDB v${MONGODB_VERSION}"
-v${MONGODB_DIR}/bin/mongod --nojournal --journalCommitInterval 500 --syncdelay 0 \
-    --wiredTigerStatisticsLogDelaySecs 0 --dbpath ${DATA_DIR} --bind_ip 127.0.0.1 &> /tmp/mongodb.log &
+${MONGODB_DIR}/bin/mongod --nojournal --journalCommitInterval 500 --syncdelay 0 \
+    --wiredTigerStatisticsLogDelaySecs 0 --noIndexBuildRetry --noscripting \
+    --dbpath ${DATA_DIR} --bind_ip 127.0.0.1 &> /tmp/mongodb.log &
 MONGODB_PID=$!
 
 # Give process some time to start up