--- a/st2common/tests/unit/test_virtualenvs.py
+++ b/st2common/tests/unit/test_virtualenvs.py
@@ -54,7 +54,8 @@ class VirtualenvUtilsTestCase(CleanFilesTestCase):
 
         # Create virtualenv
         # Note: This pack has no requirements
-        setup_pack_virtualenv(pack_name=pack_name, update=False)
+        setup_pack_virtualenv(pack_name=pack_name, update=False,
+                              include_pip=False, include_setuptools=False, include_wheel=False)
 
         # Verify that virtualenv has been created
         self.assertVirtulenvExists(pack_virtualenv_dir)
@@ -68,13 +69,15 @@ class VirtualenvUtilsTestCase(CleanFilesTestCase):
         self.assertFalse(os.path.exists(pack_virtualenv_dir))
 
         # Create virtualenv
-        setup_pack_virtualenv(pack_name=pack_name, update=False)
+        setup_pack_virtualenv(pack_name=pack_name, update=False,
+                              include_pip=False, include_setuptools=False, include_wheel=False)
 
         # Verify that virtualenv has been created
         self.assertVirtulenvExists(pack_virtualenv_dir)
 
         # Re-create virtualenv
-        setup_pack_virtualenv(pack_name=pack_name, update=False)
+        setup_pack_virtualenv(pack_name=pack_name, update=False,
+                              include_pip=False, include_setuptools=False, include_wheel=False)
 
         # Verify virtrualenv is still there
         self.assertVirtulenvExists(pack_virtualenv_dir)
@@ -88,13 +91,15 @@ class VirtualenvUtilsTestCase(CleanFilesTestCase):
         self.assertFalse(os.path.exists(pack_virtualenv_dir))
 
         # Create virtualenv
-        setup_pack_virtualenv(pack_name=pack_name, update=False)
+        setup_pack_virtualenv(pack_name=pack_name, update=False,
+                              include_pip=False, include_setuptools=False, include_wheel=False)
 
         # Verify that virtualenv has been created
         self.assertVirtulenvExists(pack_virtualenv_dir)
 
         # Update it
-        setup_pack_virtualenv(pack_name=pack_name, update=True)
+        setup_pack_virtualenv(pack_name=pack_name, update=True,
+                              include_pip=False, include_setuptools=False, include_wheel=False)
 
         # Verify virtrualenv is still there
         self.assertVirtulenvExists(pack_virtualenv_dir)