--- a/st2tests/st2tests/fixtures/packs/dummy_pack_2/requirements.txt
+++ b/st2tests/st2tests/fixtures/packs/dummy_pack_2/requirements.txt
@@ -1,1 +1,1 @@
-requests
+six