This pull request contains speed ups for various "offensive" tests which took a long time.

For information on how speed ups were achieved, please refer to the actual git commit messages and code changes (and ask for clarification, if needed).

## Speed ups

Before

```bash
(virtualenv)vagrant@vagrant-ubuntu-trusty-64:/data/stanley$ nosetests --with-timer --nocapture -v st2common/tests/unit/test_virtualenvs.py
test_setup_pack_virtualenv_already_exists (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (8.5302s)
test_setup_pack_virtualenv_doesnt_exist_yet (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (5.0522s)
test_setup_virtualenv_invalid_dependency_in_requirements_file (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (10.7242s)
test_setup_virtualenv_update (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (5.6160s)

[success] 35.84% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_virtualenv_invalid_dependency_in_requirements_file: 10.7242s
[success] 28.51% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_pack_virtualenv_already_exists: 8.5302s
[success] 18.77% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_virtualenv_update: 5.6160s
[success] 16.88% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_pack_virtualenv_doesnt_exist_yet: 5.0522s
----------------------------------------------------------------------
Ran 4 tests in 29.929s
````

After:

```bash
(virtualenv)vagrant@vagrant-ubuntu-trusty-64:/data/stanley$ nosetests --with-timer --nocapture -v st2common/tests/unit/test_virtualenvs.py
test_setup_pack_virtualenv_already_exists (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (0.2725s)
test_setup_pack_virtualenv_doesnt_exist_yet (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (0.1489s)
test_setup_virtualenv_invalid_dependency_in_requirements_file (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (3.1897s)
test_setup_virtualenv_update (tests.unit.test_virtualenvs.VirtualenvUtilsTestCase) ... ok (3.7518s)

[success] 50.96% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_virtualenv_update: 3.7518s
[success] 43.32% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_virtualenv_invalid_dependency_in_requirements_file: 3.1897s
[success] 3.70% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_pack_virtualenv_already_exists: 0.2725s
[success] 2.02% tests.unit.test_virtualenvs.VirtualenvUtilsTestCase.test_setup_pack_virtualenv_doesnt_exist_yet: 0.1489s
----------------------------------------------------------------------
Ran 4 tests in 7.368s

OK
```